requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6
Babel>=1.3
Flask>=0.10,<1.0
keystonemiddleware>=4.0.0
oslo.log>=1.12.0  # Apache-2.0
oslo.middleware>=3.0.0                  # Apache-2.0
python-heatclient>=0.6.0
python-keystoneclient>=1.6.0,!=1.8.0
python-swiftclient>=2.2.0
